### 2019-03-20 #1339 @whedon
[PRE REVIEW]: prysm: A Python optics module
**Submitting author:** @brandondube (<a href="http://orcid.org/0000-0002-3837-605X">Brandon Dube</a>)
**Repository:** <a href="https://github.com/brandondube/prysm" target ="_blank">https://github.com/brandondube/prysm</a>
**Version:** v0.15.1
