### 2019-03-20 #1339 @arfon
@brandondube - please either rename `references.bib` to be `paper.bib` or change the entry in your `paper.md` to point to `references.bib`
