### 2019-03-24 9a9e58e
update paper references
