+++ paper/paper.bib
@@ -35,7 +35,7 @@
   publisher = {PeerJ Inc.},
 }
 
-@Book{collette_python_hdf5_2014,
+@Book{hdf5pythonbook,
   title     = {Python and HDF5},
   publisher = {O'Reilly},
   year      = {2013},
@@ -67,6 +67,20 @@
   year      = {2010},
   editor    = {St\'efan van der Walt and Jarrod Millman},
   pages     = {51 - 56},
+}
+
+@Book{numpyguide,
+  title  = {A guide to NumPy},
+  year   = {2006},
+  author = {Oliphant, Travis E.},
+  volume = {1},
+}
+
+@Misc{scipy,
+  author = {Eric Jones and Travis Oliphant and Pearu Peterson and others},
+  title  = {{SciPy}: Open source scientific tools for {Python}},
+  year   = {2001},
+  url    = {http://www.scipy.org/},
 }
 
 @Comment{jabref-meta: databaseType:bibtex;}
