+++ docs/source/api/io.rst
@@ -1,0 +1,6 @@
+********
+prysm.io
+********
+
+.. automodule:: prysm.io
+    :members:
