+++ docs/source/api/convolution.rst
@@ -1,0 +1,6 @@
+*****************
+prysm.convolution
+*****************
+
+.. automodule:: prysm.convolution
+    :members:
