+++ docs/source/api/interferogram.rst
@@ -1,0 +1,6 @@
+*******************
+prysm.interferogram
+*******************
+
+.. automodule:: prysm.interferogram
+    :members:
