+++ docs/source/api/otf.rst
@@ -1,0 +1,6 @@
+*********
+prysm.otf
+*********
+
+.. automodule:: prysm.otf
+    :members:
