+++ docs/source/api/util.rst
@@ -1,0 +1,6 @@
+**********
+prysm.util
+**********
+
+.. automodule:: prysm.util
+    :members:
