+++ docs/source/api/pupil.rst
@@ -1,0 +1,6 @@
+***********
+prysm.pupil
+***********
+
+.. automodule:: prysm.pupil
+    :members:
