+++ docs/source/api/index.rst
@@ -1,0 +1,25 @@
+********
+API
+********
+
+.. toctree::
+   :maxdepth: 1
+
+   conf
+   convolution
+   coordinates
+   detector
+   fttools
+   geometry
+   interferogram
+   io
+   mtf_utils
+   objects
+   otf
+   propagation
+   psf
+   pupil
+   sample_data
+   thinlens
+   util
+   zernike
