+++ docs/source/api/mtf_utils.rst
@@ -1,0 +1,6 @@
+***************
+prysm.mtf_utils
+***************
+
+.. automodule:: prysm.mtf_utils
+    :members:
