+++ docs/source/api/zernike.rst
@@ -1,0 +1,6 @@
+*************
+prysm.zernike
+*************
+
+.. automodule:: prysm.zernike
+    :members:
