### 2019-03-27 b76fc30
fix bugs in support calculation and nsamples for convolutions
