### 2019-03-27 641dff9
tweak convolution guts
