+++ docs/source/api/index.rst
@@ -1,6 +1,6 @@
-********
-API
-********
+*************
+API Reference
+*************
 
 .. toctree::
    :maxdepth: 1
