+++ docs/source/releases/index.rst
@@ -1,0 +1,8 @@
+***************
+Release History
+***************
+
+.. toctree::
+   :maxdepth: 1
+
+   v0.15
