### 2019-03-27 00c0d3d
partial port of v0.15 release notes
