+++ docs/source/releases/v0.15.rst
@@ -1,0 +1,109 @@
+***********
+prysm v0.15
+***********
+
+New Features
+============
+
+* Surface/Wavefront error synthesis: :mod:`~prysm.interferogram` now contains the :func:`~prysm.interferogram.synthesize_surface_from_psd` core method and :func:`~prysm.interferogram.render_synthetic_surface` and :meth:`~prysm.interferogram.Interferogram.render_from_psd` convenience wrappers for synthesizing surface or wavefront data from PSD curves.  Examples of this technique can be seen in e.g. _E. Sidick Power Spectral Density Specification and Analysis of Large Optical Surfaces_.
+
+* convenience wrapper :meth:`~prysm.interferogram.Interferogram.fit_zernikes` replacing :code:`zernikefit(i.phase, ...)` invocation.
+
+* :func:`~prysm.io.write_zygo_ascii` function in :mod:`prysm.io` to write Zygo ASCII files.
+
+* :meth:`~prysm.interferogram.Interferogram.save_zygo_ascii` to write an interferogram to Zygo ASCII format.
+
+* :code:`zorder` parameter in line-based plotting functions -- :meth:`prysm._phase.OpticalPhase.plot_slice_xy`, :meth:`prysm.convolution.Convolvable.plot_slice_xy`, :meth:`prysm.interferogram.Interferogram.plot_psd_xy_avg`
+
+* :code:`mode` argument on :meth:`prysm.interferogram.Interferogram.plot_psd_xy_avg` to switch between x axis units of spatial frequency (:code:`mode='freq'`) or spatial period (:code:`mode='period'`).
+
+* `Interferogram.psd_slices` and `Interferogram.plot_psd_slices` methods replacing `psd_xy_avg` method.  Two new inquiries are `azmin` and `azmax` for the azimuthal minimum and azimuthal maximum.
+
+* `PSF.polychromatic` staticmethod to create polychromatic PSFs from ensembles of monochromatic ones.  This essentially reintroduces the `MultispectralPSF` class's functionality from earlier versions of prysm.
+
+* more configuration options.  `prysm.config` now has parameters for `Q`, `phase_colormap`, `image_colormap`, `lw`, `zorder` for controlling the default values of these parameters throughout the library.
+
+* new constants in `prysm.psf` -- `FIRST_AIRY_ZERO`, `SECOND_AIRY_ZERO`, AND `THIRD_AIRY_ZERO` as well as `SECOND_AIRY_ENCIRCLED` AND `THIRD_AIRY_ENCIRCLED`.  These concern the zeros of the airy disk and how much of the total energy is contained within.  They are all wrapped in `AIRYDATA`, a dictionary with keys of 1,2,3 and values that are length-2 tuples of (radius, encircled energy).
+
+Beta Features
+=============
+
+* `prysm.otf.long_exposure_otf` and `prysm.otf.estimate_Cn` for calculating the OTF (MTF) associated with a 'long' exposure through atmospheric turbulence.  Note that while the equations have been implemented, the results have not been checked against published values.  Please provide feedback.
+
+Improved Packaging
+==================
+
+* prysm now uses `setup.cfg` and some setuptools tricks.  It now has the `prysm.__version__` attribute and can be more easily scanned by crawlers without executing setup.py.
+
+Improved Documentation
+======================
+
+* The User's guide and Examples sections of the documentation are now jupyter notebooks and have embedded graphics and output.
+
+* There are several new examples.
+
+Improved Test Coverage
+======================
+
+* Test coverage is now > 80%
+
+Breaking API Changes
+====================
+
[…]
