+++ prysm/otf.py
@@ -536,6 +536,7 @@ def estimate_Cn(P=1013, T=273.15, Ct=1e-4):
         atmospheric struction constant of temperature, typically 10^-5 - 10^-2 near the surface
 
     Returns
+    -------
     `float`
         Cn
 
