### 2019-03-27 b4f8811
finish porting v0.15 docs
