+++ docs/source/releases/v0.15.rst
@@ -13,27 +13,27 @@ New Features
 
 * :meth:`~prysm.interferogram.Interferogram.save_zygo_ascii` to write an interferogram to Zygo ASCII format.
 
-* :code:`zorder` parameter in line-based plotting functions -- :meth:`prysm._phase.OpticalPhase.plot_slice_xy`, :meth:`prysm.convolution.Convolvable.plot_slice_xy`, :meth:`prysm.interferogram.Interferogram.plot_psd_xy_avg`
+* :code:`zorder` parameter in line-based plotting functions -- :meth:`prysm._phase.OpticalPhase.plot_slice_xy`, :meth:`prysm.convolution.Convolvable.plot_slice_xy`, :meth:`prysm.interferogram.Interferogram.plot_psd_slices`
 
-* :code:`mode` argument on :meth:`prysm.interferogram.Interferogram.plot_psd_xy_avg` to switch between x axis units of spatial frequency (:code:`mode='freq'`) or spatial period (:code:`mode='period'`).
+* :code:`mode` argument on :meth:`prysm.interferogram.Interferogram.plot_psd_slices` to switch between x axis units of spatial frequency (:code:`mode='freq'`) or spatial period (:code:`mode='period'`).
 
-* `Interferogram.psd_slices` and `Interferogram.plot_psd_slices` methods replacing `psd_xy_avg` method.  Two new inquiries are `azmin` and `azmax` for the azimuthal minimum and azimuthal maximum.
+* :meth:`prysm.interferogram.Interferogram.psd_slices` and :meth:`prysm.interferogram.Interferogram.plot_psd_slices` methods replacing `psd_xy_avg` method.  Two new inquiries are :code:`azmin` and :code:`azmax` for the azimuthal minimum and azimuthal maximum.
 
-* `PSF.polychromatic` staticmethod to create polychromatic PSFs from ensembles of monochromatic ones.  This essentially reintroduces the `MultispectralPSF` class's functionality from earlier versions of prysm.
+* :meth:`prysm.psf.PSF.polychromatic` staticmethod to create polychromatic PSFs from ensembles of monochromatic ones.  This essentially reintroduces the `MultispectralPSF` class's functionality from earlier versions of prysm.
 
-* more configuration options.  `prysm.config` now has parameters for `Q`, `phase_colormap`, `image_colormap`, `lw`, `zorder` for controlling the default values of these parameters throughout the library.
+* more configuration options.  :data:`~prysm.conf.config` now has parameters for :code:`Q`, :code:`phase_colormap`, :code:`image_colormap`, :code:`lw`, :code:`zorder` for controlling the default values of these parameters throughout the library.
 
-* new constants in `prysm.psf` -- `FIRST_AIRY_ZERO`, `SECOND_AIRY_ZERO`, AND `THIRD_AIRY_ZERO` as well as `SECOND_AIRY_ENCIRCLED` AND `THIRD_AIRY_ENCIRCLED`.  These concern the zeros of the airy disk and how much of the total energy is contained within.  They are all wrapped in `AIRYDATA`, a dictionary with keys of 1,2,3 and values that are length-2 tuples of (radius, encircled energy).
[…]
