### 2019-03-27 797f8fe
harden read_mtf_vs_field for windows
