+++ docs/source/releases/v0.14.rst
@@ -1,0 +1,107 @@
+***********
+prysm v0.14
+***********
+
+Version 0.14 introduces a host of new features and critical improvements to existing features of prysm.  Users are encouraged to upgrade from prior releases.
+
+With version 0.15, work will continue on improving the documentation and tests.  When documentation becomes "complete" and coverage exceeds 90%, version 1.0 will be released and prysm will follow more typical semver release patterns.
+
+## New Features
+
+* :func:`~prysm.fttools.pad2d` from :mod:`prysm.fttools` now takes the :code:`mode` kwarg, wrapping `numpy.pad <https://docs.scipy.org/doc/numpy-1.15.1/reference/generated/numpy.pad.html>`_ in the non-constant case.
+
+* :func:`~prysm.propagation.prop_pupil_plane_to_psf_plane` now takes the :code:`incoherent` (default :code:`True`) argument.  When :code:`incoherent=False`, the (complex-valued) coherent impulse response is returned.
+
+* wrap-around effects in convolutions have been reduced.
+
+* there is a new :func:`~prysm.geometry.truecircle` mask in :mod:`prysm.geometry` which has anti-aliased edges for improved simulation accuracy.
+
+* :func:`~prysm.io.read_mtfmapper_sfr_single` function in :mod:`prysm.io` to read outputs from `MTF Mapper <http://mtfmapper.blogspot.com/>`_ with the :code:`-f --single-roi` arguments.
+
+* :attr:`semidiameter` attr on :class:`~prysm._phase.OpticalPhase` class and subclasses (:class:`FringeZernike`, :class:`Interferogram`, ...).
+
+* :code:`show_colorbar` option on :meth:`~prysm._phase.OpticalPhase.plot2d`.
+
+* all masks in :mod:`prysm.geometry` now take a :code:`radius` argument.
+
+* :meth:`prysm.interferogram.Interferogram.mask` now takes descriptive arguments, e.g. :code:`i.mask('circle', diameter=100)` for a 100mm diameter circle.  The :code:`mask` kwarg still exists for user-provided masks.
+
+* :attr:`prysm.interferogram.Interferogram.pvr` for PVr analysis.
+
+* in :mod:`prysm.fringezernike`: :func:`fzname` function to return the name of the nth Fringe Zernike with :code:`base` (0 or 1).
+
+* :func:`fzset_to_magnitude_angle` function to convert a list of (X-Y) Zernikes to (magnitude-angle) form.
+
+* :attr:`FringeZernike.magnitudes` property to access :func:`fzset_to_magnitude_angle` on a :class:`FringeZernike` instance.
+
+* :meth:`~prysm.fringezernike.FringeZernike.top_n` method for :class:`FringeZernike` pupils to list the top n coefficients by magnitude.
+
+* :meth:`~prysm.fringezernike.FringeZernike.barplot` method for :class:`FringeZernike` pupils to plot their coefficients.
+
+* :meth:`~prysm.fringezernike.FringeZernike.barplot_magnitudes` method to plot their pairwise magnitudes (e.g, one bar for primary astigmatism).
+
+* :meth:`~prysm.fringezernike.FringeZernike.barplot_topn` method to plot the top n coefficients only.
+
+* :meth:`~prysm.fringezernike.FringeZernike.truncate` method to reduce :class:`FringeZernike` pupils to the first n terms.
+
+* :meth:`~prysm.fringezernike.FringeZernike.truncate_topn` method to reduce to top n terms.
+
+* :attr:`~prysm.detector.Detector.fs` and :attr:`~prysm.detector.Detector.nyquist` properties on the :class:`~prysm.detector.Detector` class for the sampling and nyquist frequencies in cy/mm.
+
[…]
