+++ docs/source/releases/index.rst
@@ -6,3 +6,4 @@ Release History
    :maxdepth: 1
 
    v0.15
+   v0.14
