+++ docs/source/releases/index.rst
@@ -5,5 +5,6 @@ Release History
 .. toctree::
    :maxdepth: 1
 
+   v0.16
    v0.15
    v0.14
