### 2019-03-28 4c8cca5
add examples and user's guide to readme
