### 2019-03-28 05753c9
adjust autodoc defaults
