+++ docs/source/api/index.rst
@@ -5,6 +5,7 @@ API Reference
 .. toctree::
    :maxdepth: 1
 
+   base_classes
    conf
    convolution
    coordinates
