### 2019-03-29 #1352 @whedon
[REVIEW]: prysm: A Python optics module
**Submitting author:** @brandondube (<a href="http://orcid.org/0000-0002-3837-605X">Brandon Dube</a>)
**Repository:** <a href="https://github.com/brandondube/prysm" target ="_blank">https://github.com/brandondube/prysm</a>
**Version:** v0.16
