### 2019-03-29 a1b7371
add SigFit to docs
