### 2019-03-29 e2e0f5e
lint, refactor SigFit extraction
