### 2019-03-29 8aec416
+ sigfit rigid body parser
