### 2019-03-30 ca2eb80
add sigfit rigidbody to changelog
