### 2019-03-30 e385101
add postprocess chain to convolution methods
