+++ prysm/convolution.py
@@ -93,13 +93,19 @@ class Convolvable(BasicData):
         ax.legend(title='Slice', loc='upper right')
         return fig, ax
 
-    def conv(self, other, dryrun=False):
+    def conv(self, other, postprocess=(abs,), dryrun=False):
         """Convolves this convolvable with another.
 
         Parameters
         ----------
         other : `Convolvable`
             A convolvable object
+        postprocess: `iterable` or callable or None, optional
+            function or sequence of functions to call to post-process the result.
+            None if no work should be done.  Functions should accept a single
+            argument, the spatial representation of the result, centered on the origin
+        dryrun : `bool`, optional
+            if True, return a string describing the process of convolution for this case
 
         Returns
         -------
@@ -155,7 +161,7 @@ class Convolvable(BasicData):
                 correct.
 
         """
-        return _conv(self, other, dryrun=dryrun)
+        return _conv(self, other, postprocess=postprocess, dryrun=dryrun)
 
     def deconv(self, other, balance=1000, reg=None, is_real=True, clip=False, postnormalize=True):
         """Perform the deconvolution of this convolvable object by another.
@@ -365,14 +371,14 @@ class Convolvable(BasicData):
                            unit_x=ux, unit_y=uy, has_analytic_ft=False)
 
 
-def _conv(convolvable1, convolvable2, dryrun=False):
+def _conv(convolvable1, convolvable2, postprocess=(abs,), dryrun=False):
     if convolvable1.has_analytic_ft and convolvable2.has_analytic_ft:
         if dryrun:
             return """Double analytical convolution:
                       Result is a new Convolvable with .analytic_ft that calls
                       both of the "parent" convolvables' .analytic_ft methods."""
         else:
-            return double_analytical_ft_convolution(convolvable1, convolvable2)
+            return double_analytical_ft_convolution(convolvable1, convolvable2, postprocess)
     elif convolvable1.has_analytic_ft and not convolvable2.has_analytic_ft:
         if dryrun:
             header = 'Single analytical convolution, result computed using:'
@@ -380,7 +386,7 @@ def _conv(convolvable1, convolvable2, dryrun=False):
             line2 = f'and analyical equation from {convolvable1}.'
             return '\n'.join([header, line1, line2])
         else:
-            return single_analytical_ft_convolution(convolvable2, convolvable1)
+            return single_analytical_ft_convolution(convolvable2, convolvable1, postprocess)
     elif not convolvable1.has_analytic_ft and convolvable2.has_analytic_ft:
         if dryrun:  # could refactor the duplicate of header/line1/line2
             header = 'Single analytical convolution, result computed using:'
@@ -388,7 +394,7 @@ def _conv(convolvable1, convolvable2, dryrun=False):
             line2 = f'and analyical equation from {convolvable2}.'
             return '\n'.join([header, line1, line2])
         else:
-            return single_analytical_ft_convolution(convolvable1, convolvable2)
+            return single_analytical_ft_convolution(convolvable1, convolvable2, postprocess)
     else:
         if dryrun:
             header = 'numerical convolution:'
@@ -403,14 +409,22 @@ def _conv(convolvable1, convolvable2, dryrun=False):
             line6 = f'\t{output_spacing:.3f} μm sample spacing'
             return '\n'.join([header, line1, line2, line3, line4, line5, line6])
         else:
-            return pure_numerical_ft_convolution(convolvable1, convolvable2)
[…]
