### 2019-04-04 e01390b
begin reworking detector sampling and convolutional model
