### 2019-04-04 0923184
fix mislabeling on plot_psd_slices with non-mm spatial units
