### 2019-04-10 ba059ae
partial spec for convolution engine
