### 2019-04-12 1be3c25
fix normalization bug in Zernike
