### 2019-04-12 0d3aa51
document changes from PR #4
