+++ docs/source/releases/v0.16.rst
@@ -20,6 +20,10 @@ Bugfixes
 
 * :meth:`~prysm.convolution.Convolvable.conv` now produces the correct number of output samples in all cases.  Fixes #3.
 
+* unit changes have been corrected - prior results were incorrect
+
+* the :code:`norm` kwarg has improved behavior for Zernike classes, no longer setting :code:`z.normalize = True` when the :code:`norm=False` kwarg is passed.
+
 Under-the-hood Changes
 ======================
 
