### 2019-04-15 df59120
fix mask adoption in pupil mask function
