### 2019-04-15 5bdf026
accelerate propagation with Q=1
