### 2019-04-15 d48f27e
fix uint8 typo in imread
