### 2019-04-15 c3e9da9
update v0.16 release notes
