### 2019-04-16 56acef4
fix units displayed on PSD slices plots
