+++ prysm/interferogram.py
@@ -932,7 +932,7 @@ class Interferogram(OpticalPhase):
             ax.legend(title='Slice')
 
         ax.set(xlim=xlim, xlabel=f'Spatial {label} [{unit}]',
-               ylim=ylim, ylabel=r'PSD [nm²/' + f'(cy/{self.spatial_unit})²]')
+               ylim=ylim, ylabel=r'PSD [' + f'{self.phase_unit}²/(cy/{self.spatial_unit})²]')
 
         return fig, ax
 
