### 2019-04-16 22233aa
default to 1st order filters -- still strong out-of-band rejection & less group delay
