### 2019-04-21 897748b
init on basicdata to avoid linting errors
