### 2019-04-22 fb5d285
+ a few docstrings
