### 2019-04-22 4633f40
document new changes, some dtype conformance
