+++ docs/source/releases/v0.16.rst
@@ -15,6 +15,8 @@ New Features
 
 * :func:`prysm.io.read_sigfit_rigidbody` function to read rigid body motion and radius error coefficients from SigFit :code:`SUM2` files.
 
+* :meth:`prysm.interferogram.Interferogram.pad` function to pad interferograms; useful for dealing with group delay from spatial filtering.
+
 Bugfixes
 ========
 
@@ -28,3 +30,7 @@ Under-the-hood Changes
 ======================
 
 * :attr:`prysm.pupil.Pupil.strehl` now uses a more accurate formula.
+
+* :func:`prysm.propagation.prop_pupil_to_psf_plane` performance has been improved when Q=1.
+
+* some functions have had their conformance with :attr:`prysm.config.precision` improved.
