### 2019-04-23 a39df08
unit_x, unit_y => x, y && new Detector model (in progress) && some docstring revisions
