+++ prysm/convolution.py
@@ -12,25 +12,23 @@ class Convolvable(BasicData):
     """A base class for convolvable objects to inherit from."""
     _data_attr = 'data'
 
-    def __init__(self, data, unit_x, unit_y, has_analytic_ft=False):
+    def __init__(self, x, y, data, has_analytic_ft=False):
         """Create a new Convolvable object.
 
         Parameters
         ----------
+        x : `numpy.ndarray`
+            1D ndarray defining x data grid
+        y : `numpy.ndarray`
+            1D ndarray defining y data grid
         data : `numpy.ndarray`
             2D ndarray of data
-        unit_x : `numpy.ndarray`
-            1D ndarray defining x data grid
-        unit_y : `numpy.ndarray`
-            1D ndarray defining y data grid
         has_analytic_ft : `bool`, optional
             Whether this convolvable overrides self.analytic_ft, and has a known
             analytical fourier tansform
 
         """
-        self.data = data
-        self.unit_x = unit_x
-        self.unit_y = unit_y
+        self.x, self.y, self.data = x, y, data
         self.has_analytic_ft = has_analytic_ft
 
     def __str__(self):
@@ -199,7 +197,7 @@ class Convolvable(BasicData):
         if postnormalize:
             result += result.min()
             result /= result.max()
-        return Convolvable(result, self.unit_x, self.unit_y, False)
+        return Convolvable(result, self.x, self.y, False)
 
     def renorm(self):
         """Renormalize so that the peak is at a value of unity."""
@@ -236,9 +234,9 @@ class Convolvable(BasicData):
         """
         from matplotlib import colors
 
-        if self.unit_x is not None:
-            extx = [self.unit_x[0], self.unit_x[-1]]
-            exty = [self.unit_y[0], self.unit_y[-1]]
+        if self.x is not None:
+            extx = [self.x[0], self.x[-1]]
+            exty = [self.y[0], self.y[-1]]
             ext = [*extx, *exty]
         else:
             ext = None
@@ -300,7 +298,7 @@ class Convolvable(BasicData):
                 if freq_x is None or freq_y is None:
                     raise ValueError('Convolvable has analytic FT and no data, must provide x and y coordinates')
             else:
-                lx, ly, ss = len(self.unit_x), len(self.unit_y), self.sample_spacing
+                lx, ly, ss = len(self.x), len(self.y), self.sample_spacing
                 freq_x, freq_y = forward_ft_unit(ss, lx), forward_ft_unit(ss, ly)
 
             data = self.analytic_ft(freq_x, freq_y)
@@ -368,7 +366,7 @@ class Convolvable(BasicData):
         extx, exty = (s[1] * scale) / 2, (s[0] * scale) / 2
         ux, uy = m.arange(-extx, extx, scale), m.arange(-exty, exty, scale)
         return Convolvable(data=m.flip(imgarr, axis=0).astype(config.precision),
-                           unit_x=ux, unit_y=uy, has_analytic_ft=False)
+                           x=ux, y=uy, has_analytic_ft=False)
 
 
 def _conv(convolvable1, convolvable2, postprocess=(abs,), dryrun=False):
@@ -481,12 +479,12 @@ def single_analytical_ft_convolution(without_analytic, with_analytic, postproces
     padded_data = pad2d(without_analytic.data, 2, mode='reflect')
     fourier_data = m.fft2(padded_data)
     sy, sx = padded_data.shape
-    fourier_unit_x = forward_ft_unit(without_analytic.sample_spacing, sx, shift=False)
[…]
