+++ prysm/pupil.py
@@ -64,7 +64,7 @@ class Pupil(OpticalPhase):
         else:
             # data already known
             need_to_build = False
-        super().__init__(unit_x=ux, unit_y=uy, phase=phase,
+        super().__init__(x=ux, y=uy, phase=phase,
                          wavelength=wavelength, phase_unit=opd_unit, spatial_unit='mm')
         self.xaxis_label = 'Pupil ξ'
         self.yaxis_label = 'Pupil η'
@@ -258,5 +258,5 @@ class Pupil(OpticalPhase):
 
         return Pupil(wavelength=wvl, phase=interferogram.phase,
                      opd_unit=interferogram.phase_unit,
-                     ux=interferogram.unit_x, uy=interferogram.unit_y,
+                     ux=interferogram.x, uy=interferogram.y,
                      mask=~(interferogram.phase == m.nan))

[…]
