### 2019-04-23 96f100b
+ docstrings
