### 2019-04-23 66d7228
use copies of analytic convolvables (us of execution time for less user bugs)
