### 2019-04-24 d68b397
new mathops implementation basics
