### 2019-04-24 8d45059
initial commit reworking convolution engine
