+++ prysm/convolution.py
@@ -28,7 +28,7 @@ class Convolvable(BasicData):
             analytical fourier tansform
 
         """
-        self.x, self.y, self.data = x, y, data
+        super().__init__(x=x, y=y, data=data)
         self.has_analytic_ft = has_analytic_ft
 
     def __str__(self):
@@ -91,7 +91,7 @@ class Convolvable(BasicData):
         ax.legend(title='Slice', loc='upper right')
         return fig, ax
 
-    def conv(self, other, postprocess=(abs,), dryrun=False):
+    def conv(self, other):
         """Convolves this convolvable with another.
 
         Parameters
@@ -159,7 +159,8 @@ class Convolvable(BasicData):
                 correct.
 
         """
-        return _conv(self, other, postprocess=postprocess, dryrun=dryrun)
+        e = ConvolutionEngine(self, other)
+        return e.fire()
 
     def deconv(self, other, balance=1000, reg=None, is_real=True, clip=False, postnormalize=True):
         """Perform the deconvolution of this convolvable object by another.
@@ -368,280 +369,15 @@ class Convolvable(BasicData):
         return Convolvable(data=m.flip(imgarr, axis=0).astype(config.precision),
                            x=ux, y=uy, has_analytic_ft=False)
 
-
-def _conv(convolvable1, convolvable2, postprocess=(abs,), dryrun=False):
-    if convolvable1.has_analytic_ft and convolvable2.has_analytic_ft:
-        if dryrun:
-            return """Double analytical convolution:
-                      Result is a new Convolvable with .analytic_ft that calls
-                      both of the "parent" convolvables' .analytic_ft methods."""
-        else:
-            return double_analytical_ft_convolution(convolvable1, convolvable2, postprocess)
-    elif convolvable1.has_analytic_ft and not convolvable2.has_analytic_ft:
-        if dryrun:
-            header = 'Single analytical convolution, result computed using:'
-            line1 = f'numerical data from {convolvable2}'
-            line2 = f'and analyical equation from {convolvable1}.'
-            return '\n'.join([header, line1, line2])
-        else:
-            return single_analytical_ft_convolution(convolvable2, convolvable1, postprocess)
-    elif not convolvable1.has_analytic_ft and convolvable2.has_analytic_ft:
-        if dryrun:  # could refactor the duplicate of header/line1/line2
-            header = 'Single analytical convolution, result computed using:'
-            line1 = f'numerical data from {convolvable1}'
-            line2 = f'and analyical equation from {convolvable2}.'
-            return '\n'.join([header, line1, line2])
-        else:
-            return single_analytical_ft_convolution(convolvable1, convolvable2, postprocess)
-    else:
-        if dryrun:
-            header = 'numerical convolution:'
-            line1 = f'{convolvable1}'
-            line2 = f'{convolvable2}'
-            outputwindow = max(convolvable1.support, convolvable2.support)
-            output_spacing = min(convolvable1.sample_spacing, convolvable2.sample_spacing)
-            nsamples = int(outputwindow // output_spacing) + 1
-            line3 = 'output window will have:'
-            line4 = f'\t{nsamples} samples'
-            line5 = f'\t{outputwindow:.3f} μm support'
-            line6 = f'\t{output_spacing:.3f} μm sample spacing'
-            return '\n'.join([header, line1, line2, line3, line4, line5, line6])
-        else:
-            return pure_numerical_ft_convolution(convolvable1, convolvable2, postprocess)
-
-
[…]
