+++ prysm/objects.py
@@ -268,7 +268,6 @@ class SlantedEdge(Convolvable):
         ext = samples / 2 * sample_spacing
         x = y = m.arange(-ext, ext, sample_spacing)
         xx, yy = m.meshgrid(x, y)
-        sf = samples * sample_spacing
 
         angle = m.radians(angle)
         xp = xx * m.cos(angle) - yy * m.sin(angle)
@@ -284,4 +283,4 @@ class SlantedEdge(Convolvable):
         self.contrast = contrast
         self.black = diff
         self.white = 1 - diff
-        super().__init__(data=arr, x=x * sf, y=y * sf, has_analytic_ft=False)
+        super().__init__(data=arr, x=x, y=y, has_analytic_ft=False)
