### 2019-04-24 9618c8b
document convolution engine change
