### 2019-04-24 d2ea282
+ Wavefront class, top-level export of ConvolutionEngine and Wavefront
