### 2019-04-24 5d1efd5
implement split step convolve-and-resample for detector capture
