### 2019-04-25 fa39362
fix a typo for 8 bit imsave
