### 2019-04-25 f1395e5
change behavior of pitch=None catches; always take y from x if not present; prevents subtle bugs.
