### 2019-04-25 1fca1e7
interpolate complex numbers directly; separate interpolation of magnitude and phase gives incorrect results.
