### 2019-04-27 #1352 @aquilesC
I have finished reviewing Prysm. It is a very nice package and I think it fulfills all the criteria to be published in JOSS. I have two small suggestions to improve the repository for the future:

- Contributing guidelines could be expanded. Perhaps adding an extra file in the docs, like [Pandas](https://pandas-docs.github.io/pandas-docs-travis/development/contributing.html) or [Numpy](https://www.numpy.org/devdocs/dev/index.html) did. It can be useful to create issues for missing features and linking them to the project's roadmap (the capabilities of the project board on Github are quite limited, without room for discussion, assigning people, etc.) It is also important to explain how to contribute in practical *Github* matters. For example, pull requests should be of squashed commits or should they contain the entire history?

- The API documentation appears only in the *latest* version on Readthedocs but not on the *stable* version.

The suggestions above, however, do not change the code. Therefore, @brandondube can upload the program to Zenodo and get the doi. If I'm not mistaken, after this stage, things are back in the hands of @xuanxu ?
