### 2019-04-28 154ec58
exchange old mathops for new design
todo: scipy passthrough as well
