+++ prysm/psf.py
@@ -1,7 +1,8 @@
 """A base point spread function interface."""
-from scipy import optimize, interpolate
+from scipy import optimize, interpolate, special
 
 from .conf import config
+from .mathops import engine as e, jinc
 from .coordinates import cart_to_polar
 from .util import share_fig_ax, sort_xy
 from .convolution import Convolvable
@@ -9,8 +10,6 @@ from .propagation import (
     prop_pupil_plane_to_psf_plane,
     prop_pupil_plane_to_psf_plane_units,
 )
-
-from prysm import mathops as m
 
 FIRST_AIRY_ZERO = 1.220
 SECOND_AIRY_ZERO = 2.233
@@ -98,8 +97,8 @@ class PSF(Convolvable):
         # compute MTF from the PSF
         if self._mtf is None:
             self._mtf = MTF.from_psf(self)
-            nx, ny = m.meshgrid(self._mtf.x, self._mtf.y)
-            self._nu_p = m.sqrt(nx ** 2 + ny ** 2)
+            nx, ny = e.meshgrid(self._mtf.x, self._mtf.y)
+            self._nu_p = e.sqrt(nx ** 2 + ny ** 2)
             # this is meaninglessly small and will avoid division by 0
             self._nu_p[self._nu_p == 0] = 1e-99
             self._dnx, self._dny = ny[1, 0] - ny[0, 0], nx[0, 1] - nx[0, 0]
@@ -114,7 +113,7 @@ class PSF(Convolvable):
                                                          self._dnx,
                                                          self._dny)
                 out.append(self._ee[r])
-            return m.asarray(out)
+            return e.asarray(out)
         else:
             if radius not in self._ee:
                 self._ee[radius] = _encircled_energy_core(self._mtf.data,
@@ -233,9 +232,9 @@ class PSF(Convolvable):
 
         if pix_grid is not None:
             # if pixel grid is desired, add it
-            mult = m.floor(axlim / pix_grid)
+            mult = e.floor(axlim / pix_grid)
             gmin, gmax = -mult * pix_grid, mult * pix_grid
-            pts = m.arange(gmin, gmax, pix_grid)
+            pts = e.arange(gmin, gmax, pix_grid)
             ax.set_yticks(pts, minor=True)
             ax.set_xticks(pts, minor=True)
             ax.yaxis.grid(True, which='minor', color='white', alpha=0.25)
@@ -294,7 +293,7 @@ class PSF(Convolvable):
         elif axlim is 0:
             raise ValueError('computing from 0 to 0 is stupid')
         else:
-            xx = m.linspace(1e-5, axlim, npts)
+            xx = e.linspace(1e-5, axlim, npts)
             yy = self.encircled_energy(xx)
 
         fig, ax = share_fig_ax(fig, ax)
@@ -396,13 +395,13 @@ class PSF(Convolvable):
                 ref_samples_x = psf.samples_x
                 ref_samples_y = psf.samples_y
 
-        merge_data = m.zeros((ref_samples_x, ref_samples_y, len(psfs)))
+        merge_data = e.zeros((ref_samples_x, ref_samples_y, len(psfs)))
         for idx, psf in enumerate(psfs):
             # don't do anything to the reference PSF besides spectral scaling
             if idx is ref_idx:
                 merge_data[:, :, idx] = psf.data * spectral_weights[idx]
             else:
-                xv, yv = m.meshgrid(ref_x, ref_y)
+                xv, yv = e.meshgrid(ref_x, ref_y)
[…]
