### 2019-04-28 a8fab16
+ some documentation for v0.16
