### 2019-04-28 3bd4041
some typos in release docs
