### 2019-04-29 #1352 @xuanxu
@brandondube I just found a minor typo in the article proof: `propgations` -> `propagations`. Other than that everything looks good for publication.

Once you fix the typo, please create an archive in zenodo/figshare/other and report the DOI here (if needed edit the deposit metadata so it matches the title and author name of [the JOSS paper](https://github.com/openjournals/joss-papers/blob/joss.01352/joss.01352/10.21105.joss.01352.pdf))
