### 2019-04-29 60141ad
optional numbering, offset, and width on zernike barplots
