### 2019-04-29 50835f5
use np obj directly for backend, not strings
