### 2019-04-29 #1352 @brandondube
@xuanxu Thanks for the catch!  I fixed the typo in the paper.  For Zenodo/etc, do I upload prysm@v0.15.1, 0.15.1 patched with the paper typo fix, or latest?  And is the convention to zip up the entire contents of the root directory, less .git?  Or just the minimum artifacts for the software to install?
