### 2019-04-30 #1352 @xuanxu
@brandondube usually you would release a new version with all the changes made during the review process, once you do that I'll update the version here and that would be the version you need to upload to Zenodo.

The convention is to upload the zip that GitHub provides in the release page (example: [link to zip here in the 'Assets' section](https://github.com/brandondube/prysm/releases/tag/v0.15.1)). That will include the root directory you see in GitHub (no .git dir) at the moment that specific version was released.
