### 2019-04-30 f762aba
string checking on backend changes
