### 2019-04-30 f44f43f
+ Smear class to model smear from motion blur
TODO: spatial computation in init, sample spacing, etc
