### 2019-04-30 0e1eb11
fix some bugs from mathops conversion
