### 2019-04-30 57798b7
more testing tweaks, revise how mathops works with engine changes
