### 2019-04-30 ba0c0c3
fix some barh bugs
