### 2019-04-30 8bb960c
change bindown sampling accuracy requirements to be 1% instead of buggy 1 nm
