+++ docs/source/api/degredations.rst
@@ -1,0 +1,6 @@
+*****************
+prysm.degredations
+*****************
+
+.. automodule:: prysm.degredations
+    :members:
