### 2019-04-30 02d0954
minor linting
