### 2019-05-01 89fe401
bugfix floor for nsamples on convolution engines;
occasionally would result in nsample mismatch between analytic and direct
