### 2019-05-01 8b7fb9d
rework detector capture for noninteger sample spacing mismatches
accuracy is reasonably close to integer case (within 10-20%)

this was not the case with the previous algorithm
