### 2019-05-01 #1352 @brandondube
@xuanxu thanks.  The change was pushed, but to dev.  It is on master now, along with about 90 commits of other changes :)

I'm within spitting distance (a week or so) of releasing v0.16 -- would it be alright to save the finalization here until then?
