### 2019-05-01 b38437a
add skimage to travis for detector capture tests
