+++ .travis.yml
@@ -12,6 +12,7 @@ install:
       numpy>=1.15
       scipy>=1.2
       matplotlib>=3.0
+      skimage
       imageio
       pandas
       h5py
