### 2019-05-01 ed457b5
fix skimage on travis
