+++ .travis.yml
@@ -12,7 +12,7 @@ install:
       numpy>=1.15
       scipy>=1.2
       matplotlib>=3.0
-      skimage
+      scikit-image
       imageio
       pandas
       h5py
