### 2019-05-01 #1352 @xuanxu
@brandondube I'm ok with waiting for a week for the new release. Please tell us here once you are ready for the v0.16 and I'll ping aquilesC then for a quick re-review so we are sure everything is still fine. Is that ok with you @aquilesC ?
