### 2019-05-01 e81d8ea
+ pupil add/sub tests
