### 2019-05-01 ed36894
a few detector tests
