### 2019-05-01 65982cd
use off-angle smear in test
