+++ tests/test_degredations.py
@@ -3,5 +3,5 @@ from prysm import degredations
 
 
 def test_smear():
-    sm = degredations.Smear(1, 0)
+    sm = degredations.Smear(1, 1)
     assert sm.analytic_ft(0, 0) == 1 / sm.width
