### 2019-05-01 4253b1e
+ last [frame] property on detectors, two new detector tests
