### 2019-05-01 2119660
update release notes
