### 2019-05-01 be2a538
add unit_x and unit_y properties to BasicData for deprecation
