### 2019-05-01 52b2111
+ mathops API docs
