+++ docs/source/api/mathops.rst
@@ -1,0 +1,6 @@
+**********
+prysm.mathops
+**********
+
+.. automodule:: prysm.mathops
+    :members:
