### 2019-05-01 d147e0d
fix pitch setter
