### 2019-05-01 7bef0f3
fix x, y extent of convolvables from detector sampling w/o bindown
