### 2019-05-01 378e2ce
use TkAgg on detector tests so just this file can be run w/o error
