### 2019-05-01 49f69ba
+ plot azavg test
