### 2019-05-01 7ca5f63
+ PSD fitting function
