### 2019-05-01 c8d7e07
document fit_psd in release notes
