### 2019-05-01 4beebed
redo conv documentation
