### 2019-05-03 05ed841
ensure aft merging uses types.MethodTypes to avoid unbound method issues
