### 2019-05-04 f54cc20
cleanup some formatting on docs
