+++ docs/source/api/mathops.rst
@@ -1,6 +1,6 @@
-**********
+*************
 prysm.mathops
-**********
+*************
 
 .. automodule:: prysm.mathops
     :members:
