+++ docs/source/releases/index.rst
@@ -8,3 +8,4 @@ Release History
    v0.16
    v0.15
    v0.14
+   v0.13
