+++ docs/source/contributing.rst
@@ -1,0 +1,4 @@
+************
+Contributing
+************
+
