### 2019-05-04 9703c72
bugfix for GratingArray generation when > 2x2 grid
