### 2019-05-04 8f138f4
docstring for GratingArray, single-angle ability
