### 2019-05-05 d931fae
make all analytic FTs not compute meshgrids internally
