### 2019-05-05 a1dc43a
improve dtype conformance
