### 2019-05-05 9e9d572
expose radius argument for Siemens' Stars
