+++ docs/source/releases/v0.16.rst
@@ -9,7 +9,7 @@ New Features
 
 * :func:`prysm.coordinates.make_xy_grid` and :func:`~prysm.coordinates.make_rho_phi_grid` now take a :code:`radius` argument.
 
-* :meth:`prysm.objects.TiltedSquare.__init__` now takes :code:`radius` and :code:`contrast` arguments.
+* :class:`~prysm.objects.TiltedSquare`:code:`.__init__` now takes :code:`radius` and :code:`contrast` arguments.
 
 * :func:`prysm.io.read_sigfit_zernikes` function to read Zernike coefficients from `SigFit <http://sigmadyne.com/sigfit-software/>`_ :code:`OUTCOF3` files.
 
@@ -28,6 +28,8 @@ New Features
 * :class:`~prysm.objects.Grating` to synthesize grating targets.
 
 * :class:`~prysm.objects.GratingArray` to synthesize arrays of gratings.
+
+* the :code:`radius` argument is exposed on :class:`~prysm.objects.SiemensStar`:code:`.__init__`.
 
 * :func:`~prysm.interferogram.fit_psd` to fit analytic curves to PSD data.  Note that this function works best when given a reasonable guess; curve fitting extremely high dynamic range signals (such as PSDs) is not very stable.
 
