### 2019-05-05 b53d1aa
+ basic contributing guide
