### 2019-05-05 7e26966
cleanup some unit tests
