+++ tests/test_objects.py
@@ -13,7 +13,7 @@ def test_slit_renders_correctly_for_all_orientations(orientation):
 @pytest.mark.parametrize('orientation', ['h', 'v', 'crossed'])
 def test_slit_analytic_ft_correct_at_origin(orientation):
     s = objects.Slit(1, orientation=orientation)
-    aft = s.analytic_ft(0, 0)[0, 0]
+    aft = s.analytic_ft(0, 0)
     assert aft == 1 or aft == 2
 
 
@@ -24,7 +24,7 @@ def test_pinhole_renders_properly_undersized_support():
 
 def test_pinhole_analytic_ft_correct_at_origin():
     p = objects.Pinhole(1, 0, 0)
-    assert p.analytic_ft(0, 0)[0, 0] == 0.5
+    assert p.analytic_ft(0, 0) == 0.5
 
 
 @pytest.mark.parametrize('sinusoid, background', [[True, 'w'], [True, 'b'], [False, 'w'], [False, 'b']])
