### 2019-05-05 #1352 @brandondube
@aquilesC v0.16 will have a final check-over tomorrow and assuming I find nothing derailing, a release monday.  If you would like to begin your check-over, you can see the changes (summarized with git):

```bash
cd prysm
git checkout master
git pull
git diff HEAD v0.15.1 --stat
```

They are also summarized in the release notes,
https://prysm.readthedocs.io/en/latest/releases/v0.16.html
You might have to clear your cache for this to render fully up-to-date, and wait about an hour for the RTD build to complete.

I also opened several issues on the repo with some features, bugs, etc, for future releases and adding a contributing guide to the docs.

Cheers,
- Brandon
