### 2019-05-06 #1352 @aquilesC
@brandondube , everything is looking fine!
The issues and the contributing guidelines make a much clearer entry point for new contributors.
I've run the tests and the examples, and everything looks OK.

Let me know when you publish to PyPI so I test the entire install cycle, and I guess that would be it.
