### 2019-05-06 1c4254c
add ski to RTD env
