### 2019-05-06 #1352 @brandondube
The release is up on PyPI!  Once you've given the OK, I'll package it on Zenodo for the submission 🚀
