### 2019-05-08 #1352 @xuanxu
OK, everything ready for publication. Thanks for the review @aquilesC!

Pinging @openjournals/joss-eics for final acceptance
