### 2019-05-08 0fd8a83
fix window calculation;
before, welch window was severely offset, now it is properly centered
