### 2019-05-08 7c7d2b1
don't use MKL on RTD for faster builds
