+++ readthedocs-env.yml
@@ -13,5 +13,6 @@ dependencies:
   - matplotlib
   - pandas
   - imageio
+  - nomkl
   - pip:
     - .
