### 2019-05-09 #1352 @labarba
@brandondube — I'm not sure if this matters, but it looks like you pushed v0.16 as a tag, without creating the release. You can see that the "Latest release" flag is not shown next to it:
https://github.com/brandondube/prysm/releases
