### 2019-05-09 #1352 @brandondube
👋 @labarba not sure if it matters either (I uploaded the release on PyPi) - but I went ahead and turned the tag into a release on GH as well.

RTD seems to be having some widespread issues w/ conda right now, so unfortunately the docs won't be updated until those are resolved.
