### 2019-05-09 #1352 @labarba
Congratulations, @brandondube, your JOSS paper is now published? 🎉

Sincere thanks to our editor: @xuanxu, and the reviewer: @aquilesC — your contribution to JOSS is invaluable 🙏
