### 2025-03-20 #1352 @farrukhazfar
Hi Folks, wondering if this is the appropriate place to post a question - apologies if it isn't : In Wavefront.from_amp_and_phase(aperture, opd, wvl, dx) is the "aperture" neccesarily supposed to be a mask ? If it can be numbers then can it must be a "scalar" ( as in scalar diffraction theory) field and not an intensity - yes ? kind regards and thanks, Farrukh Azfar
